% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glmnet_ela}
\alias{glmnet_ela}
\title{logloss Performance for two hyperparameter.}
\format{
\code{glmnet_ela} is a data frame with 49 rows and 3 variables:
\describe{
\item{\code{logloss}}{performance meassure}
\item{\code{alpha}}{hyperparameter}
\item{\code{s}}{hyperparameter}}
}
\usage{
glmnet_ela
}
\description{
A dataset containing the \code{logloss} performance meassure and the two hyperparameter \code{s} and \code{alpha}.
The dataset is from the glmnet package.
}
\keyword{datasets}
