% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchVisHyp.R
\name{launchVisHyp}
\alias{launchVisHyp}
\title{Title Launch VisHyp}
\usage{
launchVisHyp()
}
\value{
None
}
\description{
Launch the shiny application for the VisHyp package.
}
\examples{
if (interactive()) {
   launchVisHyp()
}

}
