% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mypackage}
\alias{mypackage}
\alias{mypackage-package}
\title{mypackage rocks}
\description{
Visualizing Hyperparamer Performance Dependencies in R.
}
\author{
\strong{Maintainer}: Simon Pradel \email{simon.pradel@gmx.de}

}
