% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap}
\usage{
plotHeatmap(
  task,
  features = NULL,
  fun = NULL,
  gridsize = 20,
  scatterplot = FALSE,
  rug = TRUE,
  title = TRUE,
  numericNA = "Max"
)
}
\arguments{
\item{task}{(\link[mlr3:Task]{mlr3::Task} object) \cr
A task for which the plot should be generated.}

\item{features}{(\code{character(2)}) \cr
The names of the two columns to be plotted. If NULL, the first two feature columns will be plotted. The default is NULL.}

\item{fun}{(\code{function}) \cr
A summary function used in the rectangles of the grid. The default is \link{mean} for a regression task and
majority vote for a classification task.}

\item{gridsize}{(\code{numeric(1)}) \cr
The number of rectangles per axis. The default value is 20.}

\item{scatterplot}{(\code{logical}) \cr
If TRUE, a scatterplot will be plotted. The default is FALSE.}

\item{rug}{(\code{logical}) \cr
If TRUE, a rug will be plotted. The default is TRUE.}

\item{title}{(\code{logical}) \cr
If TRUE, the title will be plotted. The default is TRUE.}

\item{numericNA}{(\code{character(1)} \cr
If "Max", the NA values of the column are displayed as maximum value. If "Min", the NA values are displayed as
minimum value. The default is "Max".}
}
\value{
A \link{plotly} object.
}
\description{
Visualization of a heatmap from a \link[mlr3:Task]{mlr3::Task} object.
}
\examples{
library(mlr3)
data(glmnet_ela)
task_glmnet_ela = TaskRegr$new(id = "task_glmnet", backend = glmnet_ela, target = "logloss")
plotHeatmap(task_glmnet_ela)

}
\seealso{
\link{plotParallelCoordinate} \link{plotPartialDependence} \link{plotImportance}
}
