% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallelCoordinate.R
\name{plotParallelCoordinate}
\alias{plotParallelCoordinate}
\title{Parallel Coordinate Plot (PCP)}
\usage{
plotParallelCoordinate(
  task,
  features = NULL,
  labelside = "Top",
  labelangle = 0,
  colbarrange = NULL,
  constrainrange = NULL,
  labeltarget = TRUE,
  colbarreverse = FALSE,
  autosort = TRUE,
  numericNA = "Max",
  title = FALSE,
  titleheight = 0.95
)
}
\arguments{
\item{task}{(\link[mlr3:Task]{mlr3::Task} object) \cr
A task for which the plot should be generated.}

\item{features}{(\code{character}) \cr
The names of columns to be plotted. If NULL, each column will be plotted. The default is NULL.}

\item{labelside}{(\code{character(1)} \cr
The positions of the labels for the scales. It can be chosen between "Top" and "Bottom". The default is "Top".}

\item{labelangle}{(\code{numeric(1)}) \cr
The angle of the labels. The default is 0.}

\item{colbarrange}{(\code{numeric(2)}) \cr
A vector with two numeric values. Its distance specifies the range of the color bar for the labeled target.
If NULL, the color bar is displayed. The value range of the color bar starts at the minimum and ends at the
maximum of the target variable. The default is NULL.}

\item{constrainrange}{(\code{numeric(2)}) \cr
A vector with two numeric values between 0 and 1. The vector restricts each column to a range of values that lies
within the quantile of the target variable. If NULL, no restriction is plotted. The default is NULL.}

\item{labeltarget}{(\code{logical}) \cr
If TRUE, the target will be labeled. The default is TRUE.}

\item{colbarreverse}{(\code{logical}) \cr
If TRUE, the colorbar will be reversed. The default is FALSE.}

\item{autosort}{(\code{logical}) \cr
If TRUE, the columns will be ordered. The default is TRUE.}

\item{numericNA}{(\code{character(1)} \cr
If "Max", the NA values of the column are displayed as maximum value. If "Min", the NA values are displayed as minimum value.
The default is "Max".}

\item{title}{(\code{logical}) \cr
If TRUE, a title will be plotted. The default is FALSE.}

\item{titleheight}{(\code{numeric(1)}) \cr
The height of the title. The default is 0.95.}
}
\value{
A \link{plotly} object.
}
\description{
Visualization of a parallel coordinate plot from a \link[mlr3:Task]{mlr3::Task} object.
}
\examples{
library(mlr3)
data(glmnet_ela)
task_glmnet_ela = TaskRegr$new(id = "task_glmnet", backend = glmnet_ela, target = "logloss")
plotParallelCoordinate(task = task_glmnet_ela)

}
\seealso{
\link{plotHeatmap} \link{plotPartialDependence} \link{plotImportance}
}
