% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartialDependence.R
\name{plotPartialDependence}
\alias{plotPartialDependence}
\title{partial Dependence plot (PDP)

Calculate the partial Dependence for 1 or 2 Hyperparameter.}
\usage{
plotPartialDependence(task, learner, features, gridsize = 20, rug = TRUE)
}
\arguments{
\item{task}{task object}

\item{learner}{learner object}

\item{features}{the name of the Hyperparmaeter for which the partial dependence should be calculated}

\item{gridsize}{only important for 2 Hyperparameter. default is 20}

\item{rug}{logicaL: should a rug be plotted? default is FALSE}
}
\value{
ggplot object
}
\description{
partial Dependence plot (PDP)

Calculate the partial Dependence for 1 or 2 Hyperparameter.
}
