% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smashy_super}
\alias{smashy_super}
\title{logloss Performance for two hyperparameter.}
\format{
\code{smashy_super} is a data frame with 200 rows and 12 variables:
\describe{
\item{\code{yval}}{performance meassure}}
}
\usage{
smashy_super
}
\description{
A dataset containing the \code{yval} performance meassure and 11 hyperparameter.
}
\keyword{datasets}
