% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{VisHyp-package}
\alias{VisHyp}
\alias{VisHyp-package}
\title{VisHyp: Visualize Hyperparameter Performance Dependencies}
\description{
The following four different visualization options are provided in this package, which are used to visualize hyperparameter performance dependencies.
\itemize{
\item Heatmap: \link{plotHeatmap}
\item Partial Dependence Plot: \link{plotPartialDependence}
\item Importance Plot: \link{plotImportance}
\item Parallel Coordiante Plot: \link{plotParallelCoordinate}
}
}
\author{
\strong{Maintainer}: Simon Pradel \email{simon.pradel@gmx.de}

}
