% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImportance.R
\name{plotImportance}
\alias{plotImportance}
\title{Importance Plot

Calculate the Importance for each Hyperparameter.}
\usage{
plotImportance(task, learner)
}
\arguments{
\item{task}{task object}

\item{learner}{learner object}
}
\value{
ggplot object
}
\description{
Importance Plot

Calculate the Importance for each Hyperparameter.
}
\examples{
plotImportance(task, learner)

}
