% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImportance.R
\name{plotImportance}
\alias{plotImportance}
\title{Importance Plot}
\usage{
plotImportance(task, learner = NULL, loss = "mae", title = TRUE)
}
\arguments{
\item{task}{\link[mlr3:Task]{mlr3::Task} object \cr
A task for which the plot should be generated.}

\item{learner}{\link[mlr3:Learner]{mlr3::Learner} object \cr
A learner for which the plot should be generated.
If NULL, a random forest for regression or classification will be used. The default is "regr.ranger" for regression
tasks and "classif.ranger" for classification tasks.}

\item{loss}{(\code{character(1)} \cr
The loss function to be used. It can be chosen between "ce", "f1", "mae", "mse", "rmse", "mape",
"mdae", "msle", "percent_bias", "rae", "rmse", "rmsle", "rse", "rrse", "smape". The default is "mae" for a Regression
task. For for a classification task only "ce" is available.}

\item{title}{(\code{logical}) \cr
If TRUE, the title will be plotted. The default is TRUE.}
}
\value{
A \link{plotly} object.
}
\description{
Visualization of an importance plot from a \link[mlr3:Task]{mlr3::Task} object and a \link[mlr3:Learner]{mlr3::Learner}.
}
\examples{
library(mlr3)
data(glmnet_ela)
task_glmnet_ela = TaskRegr$new(id = "task_glmnet", backend = glmnet_ela, target = "logloss")
plotImportance(task_glmnet_ela)

}
\seealso{
\link{plotParallelCoordinate} \link{plotHeatmap} \link{plotPartialDependence}
}
