% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParallelCoordinate.R
\name{plotParallelCoordinate}
\alias{plotParallelCoordinate}
\title{Parallel Coordinate plot}
\usage{
plotParallelCoordinate(
  task,
  labeltarget = TRUE,
  labelside = "top",
  labelangel = 0,
  colbarrange = NULL,
  colbarreverse = FALSE,
  constrainrange = NULL,
  autoorder = TRUE
)
}
\description{
Parallel Coordinate plot
}
