% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartialDependence.R
\name{plotPartialDependence}
\alias{plotPartialDependence}
\title{Partial Dependence Plot (PDP)}
\usage{
plotPartialDependence(
  task,
  features = NULL,
  learner = NULL,
  gridsize = 20,
  rug = TRUE,
  plotICE = TRUE,
  title = TRUE
)
}
\arguments{
\item{task}{\link[mlr3:Task]{mlr3::Task} object \cr
A task for which the plot should be generated.}

\item{features}{(\code{character(1)} | \code{character(2)}) \cr
The name of the features for which the partial dependence should be calculated.
If no features are specified, the first 2 features of the Task object will be plotted. Default is NULL.}

\item{learner}{\link[mlr3:Learner]{mlr3::Learner} object \cr
A Learner for which the plot should be generated.
If NUll, then a Random Forest for regression or classification will be used. Default is NULL.}

\item{gridsize}{(\code{numeric(1)}) \cr
The number of rectangles/ sections per axis. Default is 20.}

\item{rug}{(\code{logical}) \cr
If TRUE, then a rug will be plotted. Default is TRUE.}

\item{plotICE}{(\code{logical}) \cr
If TRUE, then the individual conditional expectations (ICE) will be also plotted. Only available for one feature.
Default is TRUE.}

\item{title}{(\code{logical}) \cr
If TRUE, then a title will be plotted. Default is TRUE.}
}
\value{
A \link{plotly} object.
}
\description{
Visualize a Partial Dependence Plot from a \link[mlr3:Task]{mlr3::Task} object and a \link[mlr3:Learner]{mlr3::Learner}
}
\examples{
library(mlr3)
data(glmnet_ela)
task_glmnet_ela = TaskRegr$new(id = "task_glmnet", backend = glmnet_ela, target = "logloss")
plotPartialDependence(task_glmnet_ela)

}
\seealso{
\link{plotParallelCoordinate} \link{plotHeatmap} \link{plotImportance}
}
