% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPartialDependence.R
\name{plotPartialDependence}
\alias{plotPartialDependence}
\title{Partial Dependence Plot (PDP)}
\usage{
plotPartialDependence(
  task,
  features = NULL,
  learner = NULL,
  gridsize = 20,
  rug = TRUE,
  plotICE = TRUE,
  title = TRUE
)
}
\arguments{
\item{task}{\link[mlr3:Task]{mlr3::Task} object \cr
A task for which the plot should be generated.}

\item{features}{(\code{character(1)} | \code{character(2)}) \cr
The name of the features for which the partial dependence should be calculated.
If no features are specified, the first two features of the task object will be plotted. The default is NULL.}

\item{learner}{\link[mlr3:Learner]{mlr3::Learner} object \cr
A learner for which the plot should be generated.
If NULL, a random forest for regression or classification will be used. The default is NULL.}

\item{gridsize}{(\code{numeric(1)}) \cr
The number of rectangles/sections per axis. The default is 20.}

\item{rug}{(\code{logical}) \cr
If TRUE, a rug will be plotted. The default is TRUE.}

\item{plotICE}{(\code{logical}) \cr
If TRUE, the individual conditional expectations (ICE) will be also plotted. It is only available if only one
feature is chosen. The default is TRUE.}

\item{title}{(\code{logical}) \cr
If TRUE, the title will be plotted. The default is TRUE.}
}
\value{
A \link{plotly} object.
}
\description{
Visualization of a partial dependence plot from a \link[mlr3:Task]{mlr3::Task} object and a \link[mlr3:Learner]{mlr3::Learner}.
}
\examples{
library(mlr3)
data(glmnet_ela)
task_glmnet_ela = TaskRegr$new(id = "task_glmnet", backend = glmnet_ela, target = "logloss")
plotPartialDependence(task_glmnet_ela)

}
\seealso{
\link{plotParallelCoordinate} \link{plotHeatmap} \link{plotImportance}
}
